backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
marshmallow==3.20.1
Werkzeug==2.3.7
cloudinary==1.34.0
sendgrid==6.10.0
openai==0.28.1
celery==5.3.6
redis==4.6.0